Extract transfer list filter classes into separate files
