+++ src/gui/CMakeLists.txt
@@ -48,9 +48,6 @@ add_library(qbt_gui STATIC
     advancedsettings.h
     autoexpandabledialog.h
     banlistoptionsdialog.h
-    categoryfiltermodel.h
-    categoryfilterproxymodel.h
-    categoryfilterwidget.h
     color.h
     cookiesdialog.h
     cookiesmodel.h
@@ -101,9 +98,6 @@ add_library(qbt_gui STATIC
     speedlimitdialog.h
     statsdialog.h
     statusbar.h
-    tagfiltermodel.h
-    tagfilterproxymodel.h
-    tagfilterwidget.h
     torrentcategorydialog.h
     torrentcontentfiltermodel.h
     torrentcontentitemdelegate.h
@@ -116,6 +110,15 @@ add_library(qbt_gui STATIC
     torrentoptionsdialog.h
     trackerentriesdialog.h
     transferlistdelegate.h
+    transferlistfilters/basefilterwidget.h
+    transferlistfilters/categoryfiltermodel.h
+    transferlistfilters/categoryfilterproxymodel.h
+    transferlistfilters/categoryfilterwidget.h
+    transferlistfilters/statusfilterwidget.h
+    transferlistfilters/tagfiltermodel.h
+    transferlistfilters/tagfilterproxymodel.h
+    transferlistfilters/tagfilterwidget.h
+    transferlistfilters/trackersfilterwidget.h
     transferlistfilterswidget.h
     transferlistmodel.h
     transferlistsortmodel.h
@@ -137,9 +140,6 @@ add_library(qbt_gui STATIC
     advancedsettings.cpp
     autoexpandabledialog.cpp
     banlistoptionsdialog.cpp
-    categoryfiltermodel.cpp
-    categoryfilterproxymodel.cpp
-    categoryfilterwidget.cpp
     cookiesdialog.cpp
     cookiesmodel.cpp
     deletionconfirmationdialog.cpp
@@ -188,9 +188,6 @@ add_library(qbt_gui STATIC
     speedlimitdialog.cpp
     statsdialog.cpp
     statusbar.cpp
-    tagfiltermodel.cpp
-    tagfilterproxymodel.cpp
-    tagfilterwidget.cpp
     torrentcategorydialog.cpp
     torrentcontentfiltermodel.cpp
     torrentcontentitemdelegate.cpp
@@ -203,6 +200,15 @@ add_library(qbt_gui STATIC
     torrentoptionsdialog.cpp
     trackerentriesdialog.cpp
     transferlistdelegate.cpp
+    transferlistfilters/basefilterwidget.cpp
+    transferlistfilters/categoryfiltermodel.cpp
+    transferlistfilters/categoryfilterproxymodel.cpp
+    transferlistfilters/categoryfilterwidget.cpp
+    transferlistfilters/statusfilterwidget.cpp
+    transferlistfilters/tagfiltermodel.cpp
+    transferlistfilters/tagfilterproxymodel.cpp
+    transferlistfilters/tagfilterwidget.cpp
+    transferlistfilters/trackersfilterwidget.cpp
     transferlistfilterswidget.cpp
     transferlistmodel.cpp
     transferlistsortmodel.cpp
