+++ src/gui/gui.pri
@@ -6,9 +6,6 @@ HEADERS += \
     $$PWD/advancedsettings.h \
     $$PWD/autoexpandabledialog.h \
     $$PWD/banlistoptionsdialog.h \
-    $$PWD/categoryfiltermodel.h \
-    $$PWD/categoryfilterproxymodel.h \
-    $$PWD/categoryfilterwidget.h \
     $$PWD/color.h \
     $$PWD/cookiesdialog.h \
     $$PWD/cookiesmodel.h \
@@ -59,9 +56,6 @@ HEADERS += \
     $$PWD/speedlimitdialog.h \
     $$PWD/statsdialog.h \
     $$PWD/statusbar.h \
-    $$PWD/tagfiltermodel.h \
-    $$PWD/tagfilterproxymodel.h \
-    $$PWD/tagfilterwidget.h \
     $$PWD/torrentcategorydialog.h \
     $$PWD/torrentcontentfiltermodel.h \
     $$PWD/torrentcontentitemdelegate.h \
@@ -74,6 +68,15 @@ HEADERS += \
     $$PWD/torrentoptionsdialog.h \
     $$PWD/trackerentriesdialog.h \
     $$PWD/transferlistdelegate.h \
+    $$PWD/transferlistfilters/basefilterwidget.h \
+    $$PWD/transferlistfilters/categoryfiltermodel.h \
+    $$PWD/transferlistfilters/categoryfilterproxymodel.h \
+    $$PWD/transferlistfilters/categoryfilterwidget.h \
+    $$PWD/transferlistfilters/statusfilterwidget.h \
+    $$PWD/transferlistfilters/tagfiltermodel.h \
+    $$PWD/transferlistfilters/tagfilterproxymodel.h \
+    $$PWD/transferlistfilters/tagfilterwidget.h \
+    $$PWD/transferlistfilters/trackersfilterwidget.h \
     $$PWD/transferlistfilterswidget.h \
     $$PWD/transferlistmodel.h \
     $$PWD/transferlistsortmodel.h \
@@ -95,9 +98,6 @@ SOURCES += \
     $$PWD/advancedsettings.cpp \
     $$PWD/autoexpandabledialog.cpp \
     $$PWD/banlistoptionsdialog.cpp \
-    $$PWD/categoryfiltermodel.cpp \
-    $$PWD/categoryfilterproxymodel.cpp \
-    $$PWD/categoryfilterwidget.cpp \
     $$PWD/cookiesdialog.cpp \
     $$PWD/cookiesmodel.cpp \
     $$PWD/deletionconfirmationdialog.cpp \
@@ -146,9 +146,6 @@ SOURCES += \
     $$PWD/speedlimitdialog.cpp \
     $$PWD/statsdialog.cpp \
     $$PWD/statusbar.cpp \
-    $$PWD/tagfiltermodel.cpp \
-    $$PWD/tagfilterproxymodel.cpp \
-    $$PWD/tagfilterwidget.cpp \
     $$PWD/torrentcategorydialog.cpp \
     $$PWD/torrentcontentfiltermodel.cpp \
     $$PWD/torrentcontentitemdelegate.cpp \
@@ -161,6 +158,15 @@ SOURCES += \
     $$PWD/torrentoptionsdialog.cpp \
     $$PWD/trackerentriesdialog.cpp \
     $$PWD/transferlistdelegate.cpp \
+    $$PWD/transferlistfilters/basefilterwidget.cpp \
+    $$PWD/transferlistfilters/categoryfiltermodel.cpp \
+    $$PWD/transferlistfilters/categoryfilterproxymodel.cpp \
+    $$PWD/transferlistfilters/categoryfilterwidget.cpp \
+    $$PWD/transferlistfilters/statusfilterwidget.cpp \
+    $$PWD/transferlistfilters/tagfiltermodel.cpp \
+    $$PWD/transferlistfilters/tagfilterproxymodel.cpp \
+    $$PWD/transferlistfilters/tagfilterwidget.cpp \
+    $$PWD/transferlistfilters/trackersfilterwidget.cpp \
     $$PWD/transferlistfilterswidget.cpp \
     $$PWD/transferlistmodel.cpp \
     $$PWD/transferlistsortmodel.cpp \
