+++ src/gui/transferlistfilters/categoryfiltermodel.cpp
@@ -33,7 +33,7 @@
 
 #include "base/bittorrent/session.h"
 #include "base/global.h"
-#include "uithememanager.h"
+#include "gui/uithememanager.h"
 
 class CategoryModelItem
 {
