+++ src/gui/transferlistfilters/categoryfiltermodel.h
