+++ src/gui/transferlistfilters/categoryfilterproxymodel.cpp
