+++ src/gui/transferlistfilters/categoryfilterproxymodel.h
