+++ src/gui/transferlistfilters/categoryfilterwidget.h
