+++ src/gui/transferlistfilters/tagfiltermodel.cpp
@@ -34,7 +34,7 @@
 
 #include "base/bittorrent/session.h"
 #include "base/global.h"
-#include "uithememanager.h"
+#include "gui/uithememanager.h"
 
 namespace
 {
