+++ src/gui/transferlistfilters/tagfilterwidget.h
